Relocating MVC Model result extensions to somehwere more common and reusable
