+++ src/Tug.Server/Controllers/DscReportingController.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Tug.Messages;
+using Tug.Server.Mvc;
 using Tug.Server.Util;
 
 namespace Tug.Server.Controllers
